+++ docs/.vuepress/config.js
@@ -18,13 +18,13 @@ module.exports = {
       lang: 'en',
       title: 'Loom SDK',
       description:
-        'Loom Developer Documentation. Learn all you need to know about building your own DApps on top of Loom PlasmaChain.',
+        'Loom Developer Documentation. Learn all you need to know about building your own DApps on top of Loom Basechain.',
     },
     '/zh-CN/': {
       lang: 'zh-CN',
       title: 'Loom SDK',
       description:
-        'Loom开发文档：学习关于在Loom PlasmaChain上构建DApp所需的全部知识。',
+        'Loom开发文档：学习关于在Loom Basechain上构建DApp所需的全部知识。',
     },
     '/ko/': {
       lang: 'ko',
@@ -36,7 +36,7 @@ module.exports = {
       lang: 'ja',
       title: 'Loom SDK',
       description:
-        'Loom 開発者ドキュメント。LoomのPlasmaChain上でのDApp開発に必要な全てを学びましょう。',
+        'Loom 開発者ドキュメント。LoomのBasechain上でのDApp開発に必要な全てを学びましょう。',
     },
   },
   themeConfig: {
@@ -165,12 +165,12 @@ module.exports = {
         collapsable: true,
         sidebarDepth: 1,
         children: [
-          ['/testnet-plasma', 'PlasmaChain Testnets'],
-          ['/mainnet-plasma', 'PlasmaChain MainNet'],
+          ['/testnet-basechain', 'Basechain Testnets'],
+          ['/mainnet-basechain', 'Basechain MainNet'],
           ['/validator', 'Become Validator'],
           ['/non-validator-node', 'Non Validator Node'],
-          ['/jump-start-plasma', 'plasma Jump Start'],
-          ['/jump-start-extdev-plasma-us1', 'extdev-plasma-us1 Jump Start'],
+          ['/jump-start-plasma', 'Basechain Jump Start'],
+          ['/jump-start-extdev-plasma-us1', 'Extdev Jump Start'],
         ],
       },
       {